% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odourMatrix.R
\name{addnacols}
\alias{addnacols}
\title{Reorder odour response matrix adding nas as necessary}
\usage{
addnacols(mat, odours)
}
\arguments{
\item{mat}{odour response matrix where odours are coloumns and cells are rows}

\item{odours}{character vector of odour names}
}
\value{
gives back a tidyied up matrix with coloumns aorderd by
\code{odours} adding \code{NA}s for missing odour response data
}
\description{
given an odour response matrix with columns for different odours and rows for
different sweep timepoints, reorder columns in desired order adding NAs for
missing odours
}
