% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesubtract_heatmap.R
\name{baseline_subtract_allfreqs}
\alias{baseline_subtract_allfreqs}
\title{Subtract baseline spike rate from list of smoothed psth data}
\usage{
baseline_subtract_allfreqs(allfreqs, baselinesamples = 1:5)
}
\arguments{
\item{allfreqs}{A list with one entry per cell, where each entry contains a
matrix of smoothed psth data where rows are timepoints and columns are
odours. This is equivalent to pulling the freq component out of
\code{physplitdata::\link{smSpikes}} object.}

\item{baselinesamples}{Integer indices of the timepoints to use for the
baseline. These are 50ms time points for \code{smSpikes}}
}
\description{
Subtract baseline spike rate from list of smoothed psth data
}
\examples{
cells=c("nm20140714c1", "nm20150305c0", "nm20141128c0", "nm20140901c0")
allfreqs=lapply(physplitdata::smSpikes[cells],
function(psthsforcell) sapply(psthsforcell,function(x) x$freq))
baseline_subtract_allfreqs(allfreqs, baselinesamples=1:5)
}

