% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesubtract_heatmap.R
\name{basesubtract_heatmap_cor_dist}
\alias{basesubtract_heatmap_cor_dist}
\title{An attaempt to create a basesubtraction option first step at calculating lifetime sparsness}
\usage{
basesubtract_heatmap_cor_dist(cells, odours, col = jet.colors(20),
  basesubtract = TRUE, baselinesamples = 1:5, ...)
}
\arguments{
\item{cells}{Character specifying cells to plot OR a data.frame containing
information about those cells}

\item{odours}{Character vector specifying odours to plot}

\item{col}{Character vector of colour levels generated by a colour palette
function such as \code{\link{jet.colors}}. Defaults to
\code{jet.colors(20)}.}

\item{basesubtract}{Whether or not to subtract the baseline firing rate}

\item{baselinesamples}{Integer indices of the timepoints to use for the
baseline. These are 50ms time points for \code{smSpikes}}

\item{...}{Additional parameters passed to \code{\link{heatmap}} function}
}
\description{
An attaempt to create a basesubtraction option first step at calculating lifetime sparsness
}
