% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_array.R
\name{common_odours_for_cells}
\alias{common_odours_for_cells}
\title{Find the common set of odours for which all cells have trials}
\usage{
common_odours_for_cells(x)
}
\arguments{
\item{x}{Any identifier for the cells}
}
\value{
A character vector of odour names
}
\description{
Find the common set of odours for which all cells have trials
}
\details{
Uses \code{\link{create_raw_summary_array}} under the hood
}
\examples{
common_odours_for_cells(c("130823c0", "130809c0", "140211c0"))
}
