% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_array.R
\name{create_raw_summary_array}
\alias{create_raw_summary_array}
\title{Create the raw summary array for all spikes}
\usage{
create_raw_summary_array(x = physplitdata::smSpikes)
}
\arguments{
\item{x}{A list of smoothed psth objects for cell/odours combinations.
Defaults to \code{\link{smSpikes}}}
}
\description{
Create the raw summary array for all spikes
}
\examples{
summary_array=create_raw_summary_array()
# histogram of baseline firing for all cells / odours
hist(summary_array[,,'baseline'])
# collapse all data for different odours for the same cell
# i.e. average baseline firing rate for each cell
baseline_cell=rowMeans(summary_array[,,'baseline'], na.rm=TRUE)
hist(baseline_cell, xlab='Firing freq /Hz')

# Plot density distributions by cell group
pns=subset(PhySplitDB, Group=='PN' & cell \%in\% names(smSpikes))$cell
physplit=PhySplitDB[match(names(smSpikes), PhySplitDB$cell), ]
rownames(physplit)=physplit$cell
physplit$baseline=baseline_cell[physplit$cell]
library(ggplot2)
qplot(baseline,col=Group, data=subset(physplit, Group\%in\%c("L","O","PN")), geom='density')
}

