% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{createSummarySpikesMat}
\alias{createSummarySpikesArray}
\alias{createSummarySpikesMat}
\title{Convert raw spike summary array into matrix or array without cells/odours missing data}
\usage{
createSummarySpikesMat(summary_array, numSamplePoints = 7, NALimit = 3)

createSummarySpikesArray(summary_array, numSamplePoints = 7, NALimit = 3)
}
\arguments{
\item{summary_array}{Raw 3D array of the form generated by
\code{\link{create_raw_summary_array}}.}

\item{numSamplePoints}{The number of temporal sample points (defaults to 7)}

\item{NALimit}{The maximum number of missing responses that are acceptable}
}
\description{
\code{createSummarySpikesMat} returns a 2D matrix without
  cells/odours missing data. The matrix has \code{ncells} rows x
  (\code{nodours} * \code{numSamplePoints}) columns. Timepoints for the same
  odour are in adjacent columuns.

\code{createSummarySpikesArray} returns a 3D array without
  cells/odours missing data. The array has ncells rows x nodours columns *
  numSamplePoints slabs.
}
\details{
Some analysis (eg PCA) cannot cope with missing data. These
functions will drop cells or odours in order to give a dataset without
missing data. The choice of NALimit will affect how many cells/odours are
dropped. NALimit=3, the default for some time, is now too low since it will
result in too many odours being dropped (while keeping more cells).
}
\examples{
summary_array=create_raw_summary_array()
clean_summary_array=createSummarySpikesArray(summary_array, NALimit = 25)
clean_summary_mat=createSummarySpikesMat(summary_array, NALimit = 25)
str(clean_summary_array)
str(clean_summary_mat)
}
\seealso{
Other summary_array: \code{\link{create_raw_summary_array}}
}

