% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{createSummarySpikesMat}
\alias{createSummarySpikesMat}
\title{Given a raw spike summary array, return matrix without cells/odours missing data}
\usage{
createSummarySpikesMat(summary_array, numSamplePoints = 7, NALimit = 3)
}
\description{
uses createSummarySpikesArray
}

