% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EstimateResponseDelayByMomentMatching.R
\name{EstimateResponseDelayByMomentMatching}
\alias{EstimateResponseDelayByMomentMatching}
\title{Estimates the response delay by cross correlating the observed response with the odor profile convolved with an exponential filter.}
\usage{
EstimateResponseDelayByMomentMatching(y, x0, verbose = FALSE, plot = FALSE,
  minDelay = 0, fitOdorProfileOnError = TRUE)
}
\arguments{
\item{y}{A length T vector of the mean firing rate of the neuron in a series of bins.}

\item{x0}{A binary length T vector indicating the odor window.}

\item{verbose}{Whether to print out the results}

\item{plot}{Whether to plot the delayed exponential on top of the firing rate.}

\item{minDelay}{The minimum delay value that can be used.}

\item{fitOdorProfileOnError}{Whether to fit the odor profile itself if learning the filter parameters failed.}
}
\value{
A list consisting of

  \item{a0}{The amplitude of the exponential.}

  \item{a1}{The autoregressive coefficient of the exponential}

  \item{tau}{The time constant of the exponential. a1 = exp(-1/tau).}

  \item{l0}{The baseline firing rate.}

  \item{delay}{The delay with the highest correlation.}

  \item{status}{A string. "OK" if the parameters could be learned, "ERROR" if not.}
}
\description{
A function that detects the delay of a response relative to the
odor onset by finding the delay with maximum cross correlation of
the given response rate timecourse, with that of the odor onset
convolved with a decaying exponential. The parameters of the
exponential are found by moment matching.
}

