% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_anatomy}
\alias{heatmap_anatomy}
\title{heatmap for set of cells on nblast anatomy distance}
\usage{
heatmap_anatomy(x, col = jet.colors(20), labRow = NULL, labCol = NULL,
  ColSideColors, RowSideColors, ...)
}
\arguments{
\item{x}{A score matrix calculated by
\code{\link[nat.nblast]{nblast_allbyall}}}

\item{col}{Character vector of colour levels generated by a colour palette
function such as \code{\link{jet.colors}}. Defaults to
\code{jet.colors(20)}.}

\item{labRow}{character vectors with row and column labels to
    use; these default to \code{rownames(x)} or \code{colnames(x)},
    respectively.}

\item{labCol}{character vectors with column labels to use; defaults to cells.}

\item{ColSideColors}{(optional) character vector of length \code{ncol(x)}
    containing the color names for a horizontal side bar that may be used to
    annotate the columns of \code{x}.}

\item{RowSideColors}{(optional) character vector of length \code{nrow(x)}
    containing the color names for a vertical side bar that may be used to
    annotate the rows of \code{x}.}

\item{...}{additional arguments passed on to \code{\link{image}},
    e.g., \code{col} specifying the colors.}
}
\description{
heatmap for set of cells on nblast anatomy distance
}

