% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_cor_dist}
\alias{heatmap_cor_dist}
\title{heatmap for set of cells and odours based on correlation distance}
\usage{
heatmap_cor_dist(cells, odours, col = jet.colors(20), ...)
}
\arguments{
\item{cells}{Character specifying cells to plot OR a data.frame containing
information about those cells}

\item{odours}{Character vector specifying odours to plot}

\item{...}{Additional parameters passed to \code{\link{heatmap}} function}
}
\description{
heatmap for set of cells and odours based on correlation distance
}

