% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{jet.colors}
\alias{jet.colors}
\title{Return a colour palette function}
\usage{
jet.colors(n)
}
\arguments{
\item{n}{Number of colours in palette}
}
\description{
Return a colour palette function
}
\seealso{
\code{\link{colorRampPalette}}
}

