% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplit.analysis.R
\docType{package}
\name{physplit.analysis}
\alias{physplit.analysis}
\alias{physplit.analysis-package}
\title{Packaged and versioned analysis functions for Shahar's cells}
\description{
These functions were originally part of RandIgor.
}
\references{
Shahar Frechter, Alexander S. Bates, Sina Tootoonian,
  Michael-John Dolan, James D. Manton, Arian Jamasb, Johannes Kohl, Davi
  Bock, Gregory S.X.E. Jefferis. bioRxiv 336982
  \href{https://doi.org/10.1101/336982}{doi: 10.1101/336982}
}
