% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotcellsf}
\alias{plotcellsf}
\title{Plot spike/Vm trace for a cell using defaults appropriate for Frechter et al}
\usage{
plotcellsf(cellid, yaxis = FALSE, xaxis = FALSE, odours = NULL,
  PlotDividers = FALSE, ...)
}
\arguments{
\item{cellid}{Any id understandable by \code{\link{anyid2longid}}}

\item{yaxis}{(default TRUE)}

\item{xaxis}{(default TRUE)}

\item{odours}{Optional ordered character vector of odours (see Details)}

\item{PlotDividers}{Plot the dividing lines between odours (default TRUE)}

\item{...}{Additional arguments passed to
\code{\link[gphys]{PlotRasterFromSweeps}}}
}
\description{
Plot spike/Vm trace for a cell using defaults appropriate for Frechter et al
}
\details{
When the \code{odours} argument is specified, only the first
  presentation of any odour is kept. The odours trials will be plotted in the
  order requested. It is not possible to plot multiple trials of the same
  odour and attempting to do this by listing the same odour twice will result
  in an error.
}
\examples{
plotcellsf("130208c1", odours=c("OilBl", "E2Hex","Prpyl", "Et3HB"))
}
\seealso{
\code{\link[gphys]{PlotRasterFromSweeps}},
  \code{\link[gphys]{AddLinesToRasterPlot}},
  \code{\link[gphys]{subset.spiketimes}}, \code{\link{anyid2longid}}
}
