% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_pop_sampling.R
\name{prop.ci}
\alias{prop.ci}
\title{Approximate (1-alpha)100\% confidence interval for proportion of a population}
\usage{
prop.ci(n, pest = 0.5, N = Inf, alpha = 0.05)
}
\arguments{
\item{n}{Sample size}

\item{pest}{Estimated positive proportion}

\item{N}{Population size (default implies infinite population)}

\item{alpha}{The confidence interval is (1-alpha)*100\%}
}
\description{
Approximate (1-alpha)100\% confidence interval for proportion of a population
}
\details{
Note that this value is generally similar to that obtained with the
  sampling based approach in \code{\link{truepos_given_sample}} when
  \code{pest=0.5} but becomes an increasingly bad approximation as the
  proportion tends to 0 or 1.
}
\examples{
# 95\% confidence interval for population size 43, sample size 10 and estimated
# proportion of 0.1 ...
# expressed as a proportion
prop.ci(10, pest=0.1, N=43)
# as a number of positives
prop.ci(10, pest=0.1, N=43)*43

## Compare with sampling based calculation
prop.ci(10, pest=0.5, N=48, alpha=.1)*48
summary(truepos_given_sample(samplepos = 5, n=10, N=48))
# more different
prop.ci(10, pest=0.2, N=48, alpha=.1)*48
summary(truepos_given_sample(samplepos = 2, n=10, N=48))
}
\references{
See https://onlinecourses.science.psu.edu/stat414/node/264
}
\seealso{
Other population-sampling: \code{\link{required.sample.size}},
  \code{\link{sample_finite_population}},
  \code{\link{truepos_given_sample}}
}
