% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_pop_sampling.R
\name{sample_finite_population}
\alias{sample_finite_population}
\title{Sample from finite population with known number of true positives}
\usage{
sample_finite_population(n, N, p = NULL, npositive = NULL, replicates = 1)
}
\arguments{
\item{n}{sample size}

\item{N}{Population size}

\item{p}{positive probability (e.g. of being an LHN)}

\item{npositive}{Number of true positives (e.g. actual number LHNs in a
tract)}

\item{replicates}{Number of samples to draw}
}
\value{
A vector of length \code{replicates} with counts of sample positives.
}
\description{
Sample from finite population with known number of true positives
}
\details{
Imagine we have integers \code{1:N} (the population size). We say
  that the first \code{npositive} integers are the true positives in the
  population (npositive may be 0).

  Now we draw a sample of size \code{n} from the integers \code{1:N}. For
  this sample we say that the observed positives are those integers
  \code{<=npositive}.
}
\examples{
# Draw a random sample of size 24 tracings for a population of 96 profiles in
# a tract known to have 36 LHNs
sample_finite_population(24, N=96, npositive=36)
# Draw 1000 random samples, plot the distribution of observed sample positives
rand.samples=sample_finite_population(24, N=96, npositive=36, replicates=1000)
mean(rand.samples)
quants=quantile(rand.samples, c(0.05,0.95))
\donttest{
library(ggplot2)
qplot(rand.samples, binwidth=1, xlab='Observed Sample Positives') +
  geom_vline(xintercept = quants, colour='red')
}

\donttest{
# Compare with binomial distribution
resdf=data.frame(x=popsample(10,50,p=.5, replicates = 100000), type='popsample')
resdf=rbind(resdf, data.frame(x=rbinom(100000, size=10,p=.5), type='rbinom'))
library(ggplot2)
qplot(x, col=type, data=resdf, geom='density')
}
}
\seealso{
\code{\link{truepos_given_sample}} for estimating the true number of
  positives in a finite population given a sample.

Other population-sampling: \code{\link{prop.ci}},
  \code{\link{required.sample.size}},
  \code{\link{truepos_given_sample}}
}
