% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_pop_sampling.R
\name{truepos_given_sample}
\alias{truepos_given_sample}
\title{Estimate distribution of true positives given sampling resuts}
\usage{
truepos_given_sample(samplepos, n, N, replicates = 1000)
}
\arguments{
\item{samplepos}{Number of positives observed in sample}

\item{n}{Sample size}

\item{N}{Population size}

\item{replicates}{Number of replicates per tested true pos number}
}
\value{
a vector containing population true positive counts that could have
  generated the observed number of sample positives.
}
\description{
Estimate distribution of true positives given sampling resuts
}
\details{
The idea is to generate random realisations for all possible numbers
  of true positives, choose only those cases that resulted in the observed
  number of sample positives, and then use that empirical distribution of
  simulated true positives to estimate the unknown true positive.

  We can use this approach to estimate
}
\examples{
# Imagine we have sampled 10 profiles from a tract of 48 and found 2 LHNs
tps=truepos_given_sample(samplepos = 2, n=10, N=48)

hist(tps, breaks=0:49-.5, col='red')
plot(ecdf(tps))
# calculate median and 90\% range
quantile(tps, c(.05,.5,.95))
# calculate mode i.e. most likely true positive number
pmode <- function(x) as.integer(names(which.max(table(x))))
pmode(tps)
}
\seealso{
\code{\link{sample_finite_population}}
}
