% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{anyid2shortid}
\alias{anyid2shortid}
\title{Convert any of Shahar's ids (or a filename) to a short id}
\usage{
anyid2shortid(x)
}
\arguments{
\item{x}{A stack, cell or short id or a filename that begins with one of these.}
}
\value{
A vector of shortids, named by the original id when there were more than 1
}
\description{
Convert any of Shahar's ids (or a filename) to a short id
}
\details{
shortids look like 120726c3
}
\examples{
anyid2shortid("120726BJK1742SF274LC.lsm")
}

