% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClusterLhnData.R
\name{ClusterLhnData}
\alias{ClusterLhnData}
\title{Title}
\usage{
ClusterLhnData(X, Y, Fits, numClusters = 3, kalpha = 10, thalpha = 3/20,
  tauth = 0.1, taua = 1, tauu0 = 0.5, filterLength = 20, be = 5,
  numIters = 10000, fvalTol = 0.01, dt = 1e-05, seed = 0,
  emptyAction = "singleton", initMode = "random", verbose = TRUE,
  timer = "OFF", slopeRatioToStop = 100, numSlopePoints = 20,
  checkToStopEvery = 100)
}
\arguments{
\item{X}{dfgkdf;g asdfsadf}

\item{Y}{}

\item{Fits}{}

\item{numClusters}{}

\item{kalpha}{}

\item{thalpha}{}

\item{tauth}{}

\item{taua}{}

\item{tauu0}{}

\item{filterLength}{}

\item{be}{}

\item{numIters}{}

\item{fvalTol}{}

\item{dt}{}

\item{seed}{}

\item{emptyAction}{}

\item{initMode}{}

\item{verbose}{}

\item{timer}{}

\item{slopeRatioToStop}{}

\item{numSlopePoints}{}

\item{checkToStopEvery}{}
}
\description{
Title
}

