% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{createSummarySpikesArray}
\alias{createSummarySpikesArray}
\title{Given a raw spike summary array, drop cells/odours missing data}
\usage{
createSummarySpikesArray(summary_array, numSamplePoints = 7, NALimit = 3)
}
\description{
Given a raw spike summary array, drop cells/odours missing data
}

