% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_cor_dist}
\alias{heatmap_cor_dist}
\title{heatmap for set of cells and odours based on correlation distance}
\usage{
heatmap_cor_dist(cells, odours, col = jet.colors(20), ...)
}
\arguments{
\item{cells}{Character specifying cells to plot OR a data.frame containing
information about those cells}

\item{odours}{Character vector specifying odours to plot}

\item{col}{Character vector of colour levels generated by a colour palette
function such as \code{\link{jet.colors}}. Defaults to
\code{jet.colors(20)}.}

\item{...}{Additional parameters passed to \code{\link{heatmap}} function}
}
\description{
heatmap for set of cells and odours based on correlation distance
}
\examples{
# Find LHNs with Anatomy.type 4 and ephys class 16,17,23 for which we have
# spiking response data
physplit.c161723=subset(PhySplitDB, cell\%in\% names(Spikes) &
  Anatomy.type==4 & class\%in\%c(16,17,23))
heatmap_cor_dist(physplit.c161723)
# repeat but with class information
heatmap_cor_dist(physplit.c161723, labRow=physplit.c161723$class,
  labCol=NA, RowSideColors=rainbow(3)[factor(physplit.c161723$class)])
}

