% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physplit.analysis.R
\docType{package}
\name{physplit.analysis}
\alias{physplit.analysis}
\alias{physplit.analysis-package}
\title{Packaged and versioned Analysis functions for Shahar's cells}
\description{
these functions were originally part of RandIgor
}
